+++ src/components/sidebar/button-event-handlers.js
